This utilizes an SVG's transparent alpha, and animated linear background gradients to produce a cool effect
I had this idea to use a track matte technique from After Effects (or a
clipping mask in photoshop) to make a cool effect using just HTML and
CSS.

This utilizes a 100% wide, 100% tall SVG image with text cut out from
it. The background of the section that the image covers has an animated
gradient using CSS3.The result is a really cool effect with unlimited
creative possibilities.

You could change the size of the section to any proportion and not have
this take up 100% of the screen size.